Render harvest baskets for instanced raised beds
